extractors/api/intercom-contacts/requirements.txt
requests>=2.31.0
google-cloud-storage>=2.14.0
google-cloud-bigquery>=3.14.0
fastavro>=1.9.0